+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 Change Log
 ==========
 
+Version 0.19 *(TBD)*
+------------------
+ * Fixes ambiguous service lookup behavior of destroyed scopes:
+
+    * `MortarScope.getScope(context).isDead()` returns true when you'd expect it to.
+
+    * `ObjectGraph.inject(context, object)` throws if the backing scope is dead, as opposed to the current behavior where we instead try (and generally fail with a confusing message) to inject from an ancestor scope.
+
+    * The behavior of `MortarScope.hasService(String)` is not changed in destroyed scopes. It always says yes if the service is provided by the receiving scope or an ancestor.
+
+
 Version 0.18 *(2015-07-14)*
 ------------------
  * Destroying a scope recursively destroys its children first, like it used to.